+++ engine/source/platform/opengl/OpenGLIndexBuffer.h
@@ -7,7 +7,7 @@ namespace Cm
 class CHIMERA_API OpenGLIndexBuffer : public IndexBuffer
 {
 public:
-  OpenGLIndexBuffer( float* indices, uint32_t size );
+  OpenGLIndexBuffer( uint32_t* indices, uint32_t size );
   virtual ~OpenGLIndexBuffer();
 
   virtual void Bind() const override;
